app/monitoring/middleware.py
"""Request ID, timing, and structured logging middleware."""
import time
import uuid
from collections.abc import Callable

import structlog
from fastapi import Request, Response
from starlette.middleware.base import BaseHTTPMiddleware
from starlette.types import ASGIApp

from app.monitoring.metrics import (
    ACTIVE_REQUESTS,
    HTTP_REQUEST_DURATION_SECONDS,
    HTTP_REQUESTS_TOTAL,
)

logger = structlog.get_logger(__name__)


class RequestLoggingMiddleware(BaseHTTPMiddleware):
    """Adds request-id, timing, and structured log emission per request."""

    def __init__(self, app: ASGIApp) -> None:
        super().__init__(app)

    async def dispatch(self, request: Request, call_next: Callable) -> Response:
        request_id = str(uuid.uuid4())
        start_time = time.perf_counter()

        # Bind request context for all log calls within this request
        structlog.contextvars.clear_contextvars()
        structlog.contextvars.bind_contextvars(
            request_id=request_id,
            method=request.method,
            path=request.url.path,
        )

        request.state.request_id = request_id
        ACTIVE_REQUESTS.inc()

        try:
            response = await call_next(request)
        except Exception as exc:
            latency_ms = (time.perf_counter() - start_time) * 1000
            logger.error(
                "unhandled_exception",
                error=str(exc),
                latency_ms=round(latency_ms, 2),
            )
            raise
        finally:
            ACTIVE_REQUESTS.dec()

        latency_ms = (time.perf_counter() - start_time) * 1000
        latency_s = latency_ms / 1000

        # Prometheus labels
        endpoint = request.url.path
        HTTP_REQUESTS_TOTAL.labels(
            method=request.method,
            endpoint=endpoint,
            status_code=response.status_code,
        ).inc()
        HTTP_REQUEST_DURATION_SECONDS.labels(
            method=request.method,
            endpoint=endpoint,
        ).observe(latency_s)

        logger.info(
            "request_completed",
            status_code=response.status_code,
            latency_ms=round(latency_ms, 2),
            user_agent=request.headers.get("user-agent"),
        )

        response.headers["X-Request-ID"] = request_id
        response.headers["X-Latency-Ms"] = str(round(latency_ms, 2))
        return response
